Investment category change:
I validated successfully the following scenarios:

Created a Feature without Investment relationship, and created it with Investment Category KTLO, after the process the Investment Category returns to None

Created an Epic without Investment relationship, and created it with Investment Category KTLO, after the process the Investment Category returns to None.

Created an Epic with several Features without Investment relationship, the Epic and Features turns even if they are created with a selected Investment Category it returns to None

Changed the Investment Category of the 1st Feature to "Transformational" and it returned to "None"

Changed the Investment Category of the 1st Epic to "Transformational" and it returned to "None"

Changed the Investment Category of the Epic with children to "Transformational" and it returned to "None"

Created an Investment with Investment Category of "Short Term Revenue", added an Epic with Features. All Epic and features cascaded down the Investment Category

Moved an arphaned Epic to that Investment and Investment Category cascaded down.

For ease of use I was able to indicate the fields changes in an array,
previously every change was done individual, now if there is an update
it only validates for the CHECKED FIELDS